+++ library/src/main/java/com/wultra/android/mtokensdk/inbox/InboxService.kt
@@ -21,7 +21,6 @@ import com.wultra.android.mtokensdk.api.inbox.*
 import com.wultra.android.mtokensdk.api.inbox.model.GetList
 import com.wultra.android.mtokensdk.api.inbox.model.GetMessageDetail
 import com.wultra.android.mtokensdk.api.inbox.model.SetMessageRead
-import com.wultra.android.mtokensdk.common.Logger
 import com.wultra.android.mtokensdk.operation.OperationsUtils
 import com.wultra.android.powerauth.networking.IApiCallResponseListener
 import com.wultra.android.powerauth.networking.OkHttpBuilderInterceptor
@@ -41,8 +40,7 @@ fun PowerAuthSDK.createInboxService(appContext: Context, baseURL: String, okHttp
 fun PowerAuthSDK.createInboxService(appContext: Context, baseURL: String, strategy: SSLValidationStrategy, userAgent: UserAgent? = null): IInboxService {
     val builder = OkHttpClient.Builder()
     strategy.configure(builder)
-    Logger.configure(builder)
-    return createInboxService(appContext, baseURL, builder.build(), userAgent)
+    return InboxService(builder.build(), baseURL, this, appContext, null, userAgent)
 }
 
 class InboxService(